split user and general config
I need a place for post-config processing, and it seems to make sense to
have that in a "default" config file, portions of which can be
overridden.

This of course does render quiz.cfg kinda pointless so long as I'm doing
all my work out of a single quiz/ dir anyway, but my hindbrain has plans
for that.