quiz: add a "run control" facility and tmux support
The rc is a script (and other things) that is run by init. It is
responsible for setting up for whatever user interaction concept is
requested, then running the user program.

Two rcs are included:
- "raw": just the kernel and program output all in one, for piping or
         redirecting elsewhere
- "tmux": a pleasant "interactive" mode for the run, managed by tmux

quiz -r <runcontrol> selects the rc to use. The default is "auto", which
will chose "tmux" if output is connected to a terminal, "raw" otherwise.

The major structural change is that qemu no longer handles signals
itself. To get signal handling inside the VM, the rc must arrange a pty
allocation. Which we do, so that's cool.