quiz-prepare-kernel: fold workdir update in
The only time you need to update the workdir is after building a kernel,
and you basically don't get a choice about it. There's very little point
keeping them separate.

The new -W option is somewhat dubious, except that I neeed it to fix up
after a -K rebuild-the-world run. We'll see.