quiz-prepare-kernel: use "best" kernel image format
Raw vmlinux doesn't always have boot metadata qemu can work with (eg PVH
ELF note on 4.x kernels). So, try to use the "default" kernel
name/type/format for the architecture. We can't just do a default build
because some architectures (eg powerpc64) try to generate a number of
image variations, and we don't want to have to carry tools to support
all of them when we only need one.